supporting-projects/healthit-system-monitor/requirements.txt
psutil
colorama
python-dotenv